requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Deep learning utilities
torchvision>=0.13.0
tensorboard>=2.10.0
tqdm>=4.62.0

# Neural differential equations
torchdiffeq>=0.2.3
torchcde>=0.2.5
torchdyn>=1.0.3

# Signature computations
signatory>=1.2.6
iisignature>=0.24

# Configuration and logging
pyyaml>=5.4.0
wandb>=0.12.0
hydra-core>=1.2.0

# Scientific computing
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0